@@ -1,21 +1,19 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import { store } from './app/store';
-import { Provider } from 'react-redux';
-import * as serviceWorker from './serviceWorker';
+import ReactDOM from "react-dom/client"
+import { Provider } from "react-redux"
+import { PersistGate } from "redux-persist/integration/react"
+import { App } from "./App"
+import * as serviceWorker from "./serviceWorker"
+import { persistor, store } from "./store"
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
+root.render(
+  <Provider store={store}>
+    <PersistGate loading={null} persistor={persistor}></PersistGate>
+    <App />
+  </Provider>
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorker.unregister()
