@@ -1,58 +1,25 @@
-import React from 'react';
-import logo from './logo.svg';
-import { Counter } from './features/counter/Counter';
-import './App.css';
+import { useSelector } from "react-redux"
+import { BrowserRouter, Route, Routes } from "react-router-dom"
+import "./App.css"
+import Navbar from "./components/layout/navbar"
+import { Login } from "./pages/Login"
+import { Menu } from "./pages/Menu"
+import { Orders } from "./pages/Orders"
+import { RootState } from "./reducers"
 
-function App() {
+export const App = () => {
+  const isLoggedIn = useSelector((state: RootState) => state.isLoggedIn)
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <Counter />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <span>
-          <span>Learn </span>
-          <a
-            className="App-link"
-            href="https://reactjs.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React
-          </a>
-          <span>, </span>
-          <a
-            className="App-link"
-            href="https://redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux
-          </a>
-          <span>, </span>
-          <a
-            className="App-link"
-            href="https://redux-toolkit.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux Toolkit
-          </a>
-          ,<span> and </span>
-          <a
-            className="App-link"
-            href="https://react-redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React Redux
-          </a>
-        </span>
-      </header>
-    </div>
-  );
+    <>
+      <BrowserRouter>
+        {isLoggedIn && <Navbar /> }
+        <Routes>
+          <Route path="/" element={<Login />} />
+          {isLoggedIn && <Route path="/orders" element={<Orders />} />}
+          {isLoggedIn && <Route path="/menu" element={<Menu />} />}
+        </Routes>
+      </BrowserRouter>
+    </>
+  )
 }
-
-export default App;
